sched/eevdf: Add ENFORCE_ELIGIBILITY and default settings
Eligibility reduces scheduling latency of high nice task at the
cost of runtime of low nice tasks. This hurts the chrome model where
UI tasks should not be interrupted by background tasks. This adds
a feature to be able to disable eligibility checks.

Note that fairness is presvered in the way the deadline is calculated
since the deadline is adjusted by the weight of the task and a task.

Disabling lag placement is similar to eligibility since lag will
adjust the CPU time based on the amount of runtime.

BUG=b:308209790
TEST=Manual tested on DUT
UPSTREAM-TASK=b:324602237
Change-Id: I3598b50194e662d00c94cfba80a942aac98f6015
Signed-off-by: Youssef Esmat <youssefesmat@chromium.org>
Signed-off-by: angelomds42 <angelomds42@gmail.com>